Moved out check_payload to flaskr, and separated search by tags and id to different endpoints
